Modified display of avatar image in nav bar
I noticed that when using an avatar (gravatar, in my case) the space needed to render it in the nav bar was not "reserved" in advance. This results in getting at first the page with the nav bar without the avatar and then, once it appears, it moves the rest of the elements to the right. This was only tested on Chrome.