Add docker and helm support
This commit provides a Dockerfile based on php-fpm that contains the source
code of question2answer. The container also support init scripts that may be
used to install plugins, themes or configure the running instance.

The commit also provides a compose file which enables the possibility to run a
deploy of question2answers using docker stack.

Also includes a helm directory which enables the deploy of question2answers in
a kubernetes server using helm. For this, a public image of the Dockerfile is
needed but for now, there is none published; so you may have publish your own
until an official one is released.

Documentation about this files, how to deploy an image, a docker stack or helm
instance is also provided.